# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Blog:
    properties:
      author_id:
        description: |-
          ID of the author who wrote the blog (required)
          example: 507f1f77bcf86cd799439022
        type: string
      author_name:
        description: |-
          Author's name (required)
          example: John Doe
        type: string
      content:
        description: |-
          Blog content body (required)
          example: This is an introductory post about the Go programming language...
        type: string
      created_at:
        description: |-
          Timestamp when blog was created
          example: 2025-08-10T15:04:05Z
        type: string
      id:
        description: |-
          Unique identifier for the blog (MongoDB ObjectID)
          example: 507f1f77bcf86cd799439011
        type: string
      stats:
        allOf:
        - $ref: '#/definitions/domain.BlogStats'
        description: Engagement metrics (views, likes, etc.)
      tags:
        description: |-
          List of tags associated with the blog post (required)
          example: ["programming", "golang", "tutorial"]
        items:
          type: string
        type: array
      title:
        description: |-
          Blog post title (required)
          example: Introduction to Go
        type: string
      updated_at:
        description: |-
          Timestamp when blog was last updated
          example: 2025-08-11T16:05:06Z
        type: string
    required:
    - author_id
    - author_name
    - content
    - tags
    - title
    type: object
  domain.BlogStats:
    properties:
      comments:
        description: |-
          Number of comments on the blog
          example: 45
        type: integer
      dislikes:
        description: |-
          Number of dislikes received
          example: 10
        type: integer
      likes:
        description: |-
          Number of likes received
          example: 256
        type: integer
      views:
        description: |-
          Number of times the blog was viewed
          example: 1234
        type: integer
    type: object
...
  domain.GenerateBlogPostRequest:
    properties:
      audience:
        description: |-
          The target audience for the blog post (e.g., beginners, experts)
          example: "beginners"
        type: string
      length:
        description: |-
          The preferred length of the blog post (e.g., short, medium, long)
          example: "medium"
        type: string
      tone:
        description: |-
          The desired tone or style of the writing (e.g., formal, casual)
          example: "formal"
        type: string
      topic:
        description: |-
          The main topic or subject of the blog post
          example: "Introduction to Go concurrency"
        type: string
    type: object
  domain.ImproveBlogPostRequest:
    properties:
      content:
        description: |-
          The current content of the blog post to be improved
          example: "Go is a statically typed programming language..."
        type: string
      goal:
        description: |-
          The goal or focus of the improvement (e.g., clarity, engagement)
          example: "increase clarity"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the Blog API documentation.
  title: Blog API
  version: "1.0"
paths:
  /ai/generate:
    post:
      consumes:
      - application/json
      description: Create a blog post from a request with AI assistance
      parameters:
      - description: Blog generation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GenerateBlogPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Generate a blog post based on user input using AI
      tags:
      - AI
  /ai/improve:
    post:
      consumes:
      - application/json
      description: Enhance blog content using AI with a specified goal
      parameters:
      - description: Blog improvement request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ImproveBlogPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Improve an existing blog post based on a goal
      tags:
      - AI
  /ai/metadata:
    post:
      consumes:
      - application/json
      description: Extract title, tags, and description metadata using AI
      parameters:
      - description: Blog content
        in: body
        name: content
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Generate SEO metadata from blog content
      tags:
      - AI
  /ai/suggest:
    post:
      consumes:
      - application/json
      description: Get AI-generated suggestions to improve blog content quality and
        SEO
      parameters:
      - description: Blog content
        in: body
        name: content
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Suggest writing and SEO improvements for blog content
      tags:
      - AI
  /ai/summarize:
    post:
      consumes:
      - application/json
      description: Get a concise AI-generated summary of blog content
      parameters:
      - description: Blog content
        in: body
        name: content
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Generate a short summary of the blog content
      tags:
      - AI
...
  /blogs:
    get:
      description: Get paginated and filtered list of blog posts
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Author name
        in: query
        name: author
        type: string
      - collectionFormat: csv
        description: Tag filter
        in: query
        items:
          type: string
        name: tag
        type: array
      - description: Sort by field
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc or desc)
        in: query
        name: sort_order
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List blog posts
      tags:
      - blogs
    post:
      consumes:
      - application/json
      description: Create a new blog post by an authenticated user
      parameters:
      - description: Blog data
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/domain.Blog'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Blog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new blog post
      tags:
      - blogs
...
  /blogs/{id}:
    delete:
      description: Delete a blog post if you are the author or an admin
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a blog post
      tags:
      - blogs
    get:
      description: Retrieve a single blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Blog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a blog post by ID
      tags:
      - blogs
    put:
      consumes:
      - application/json
      description: Update an existing blog post if you are the author
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated blog data
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/domain.Blog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Blog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a blog post
      tags:
      - blogs
...
swagger: "2.0"